+++ src/Molecules/Tooltip/Tooltip.tsx
@@ -23,11 +23,11 @@ const Backdrop = styled.div`
   inset: 0 0 0 0;
   width: 100vw;
   height: 100vh;
-  background-color: ${(p) => p.theme.color.modalBackdrop};
+  background-color: ${p => p.theme.color.modalBackdrop};
   z-index: ${({ theme }) => theme.zIndex.modal};
 `;
 
-export const Tooltip: FC<Props> = (props) => {
+export const Tooltip: FC<Props> = props => {
   const {
     className,
     children,
@@ -44,7 +44,7 @@ export const Tooltip: FC<Props> = (props) => {
     wrapChild,
     pointerEvents = false,
     customBoundary,
-    pointerArrow,
+    pointerArrow = true,
     bottomSheetQuery,
     invertedColors,
   } = props;
@@ -52,7 +52,18 @@ export const Tooltip: FC<Props> = (props) => {
 
   const [triggerElement, setTriggerElement] = useState(undefined);
 
-  const bottomSheet = useMedia((t) => bottomSheetQuery?.(t) || '__false') || false;
+  const [popoverElement, setPopoverElement] = useState(undefined);
+
+  const bottomSheet = useMedia(t => bottomSheetQuery?.(t) || '__false') || false;
+
+  const params = {
+    mode: bottomSheet ? 'click' : mode,
+    controlledIsOpen,
+    openDelay,
+    closeDelay,
+    isBottomSheet: bottomSheet,
+    popoverElement,
+  };
 
   const {
     id,
@@ -65,7 +76,14 @@ export const Tooltip: FC<Props> = (props) => {
     handleMouseLeave,
     handleKeyDown,
     handleMouseDown,
-  } = useTooltip(bottomSheet ? 'click' : mode, controlledIsOpen, openDelay, closeDelay);
+  } = useTooltip(
+    params.mode,
+    params.controlledIsOpen,
+    params.openDelay,
+    params.closeDelay,
+    params.isBottomSheet,
+    params.popoverElement,
+  );
 
   return (
     <>
@@ -83,6 +101,7 @@ export const Tooltip: FC<Props> = (props) => {
       {isOpen && bottomSheet && <Backdrop />}
       {isOpen && (
         <PopOver
+          setPopoverElement={bottomSheet ? setPopoverElement : undefined}
           className={className}
           id={id}
           triggerElement={triggerElement}
@@ -93,7 +112,7 @@ export const Tooltip: FC<Props> = (props) => {
           maxWidth={maxWidth}
           offset={offset}
           pointerArrow={pointerArrow}
-          pointerEvents={pointerEvents}
+          pointerEvents={bottomSheet || pointerEvents}
           handleMouseEnter={handleMouseEnter}
           handleMouseLeave={handleMouseLeave}
           customBoundary={customBoundary}